refactor(perf-ci): extract shared runner core
