+++ development/perf-ci/lib/id.js
@@ -1,0 +1,12 @@
+function nowId() {
+  const d = new Date();
+  const pad = (n) => String(n).padStart(2, '0');
+  return `${d.getFullYear()}${pad(d.getMonth() + 1)}${pad(d.getDate())}-${pad(
+    d.getHours(),
+  )}${pad(d.getMinutes())}${pad(d.getSeconds())}`;
+}
+
+module.exports = {
+  nowId,
+};
+
